Break out lifetime.hpp
